requirements.txt
openai>=1.0.0
pyyaml>=6.0
rich>=13.0
click>=8.0
